strings: prune/simplify StringPrintf code
Remove OS and compiler compatibility code/macros and replace the core
implementation with a simple call to GNU's vasprintf function. The
original code went to great lengths to avoid using malloc'd memory but
there isn't any reason to do so here in update_engine. Additional
changes include:

 - Use GCC attributes directly.
 - Drop all support for wstring/wchar.
 - Do not simply return if printf fails, there is no way to know if
   returning an empty string to the caller is safe. Just crash...
 - Move from the `base` to `strings` namespace to avoid conflicting with
   libbase and rename files for consistency.
 - Drop SStringPrintf and StringAppendF, the former is unused in
   update_engine and the latter is easily replaced by `+=`.
